chore: remove more TODO comments
