+++ shared/types/bridge.ts
@@ -1,6 +1,7 @@
 import type { Hex } from '@metamask/utils';
 import type { BigNumber } from 'bignumber.js';
 import type { AssetType } from '../constants/transaction';
+import type { SwapsTokenObject } from '../constants/swaps';
 
 export type ChainConfiguration = {
   isActiveSrc: boolean;
@@ -183,3 +184,19 @@ export enum BridgeBackgroundAction {
   RESET_STATE = 'resetState',
   GET_BRIDGE_ERC20_ALLOWANCE = 'getBridgeERC20Allowance',
 }
+export type BridgeControllerState = {
+  bridgeFeatureFlags: BridgeFeatureFlags;
+  srcTokens: Record<string, SwapsTokenObject>;
+  srcTopAssets: { address: string }[];
+  srcTokensLoadingStatus?: RequestStatus;
+  destTokensLoadingStatus?: RequestStatus;
+  destTokens: Record<string, SwapsTokenObject>;
+  destTopAssets: { address: string }[];
+  quoteRequest: Partial<QuoteRequest>;
+  quotes: (QuoteResponse & L1GasFees)[];
+  quotesInitialLoadTime?: number;
+  quotesLastFetched?: number;
+  quotesLoadingStatus?: RequestStatus;
+  quoteFetchError?: string;
+  quotesRefreshCount: number;
+};
